Kreiran MailService projekat

+++ E-library/Security/IdentityServer/Controllers/MemberController.cs
@@ -6,6 +6,7 @@ using Microsoft.AspNetCore.Mvc;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 
 namespace IdentityServer.Controllers
@@ -36,9 +37,13 @@ namespace IdentityServer.Controllers
         [HttpPut("[action]")]
         [ProducesResponseType(StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
-        public async Task<ActionResult> Pay()
+        [ProducesResponseType(StatusCodes.Status403Forbidden)]
+        public async Task<ActionResult> Pay(string username)
         {
-            string username = Environment.UserName;
+            if (User.FindFirstValue(ClaimTypes.Name) != username)
+            {
+                return Forbid();
+            }
 
             var result = await _repository.Pay(username);
 
@@ -56,13 +61,20 @@ namespace IdentityServer.Controllers
 
         }
 
+
+
+
         [Authorize]
         [HttpDelete]
         [ProducesResponseType(StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
-        public async Task<ActionResult> DeleteAccount()
+        [ProducesResponseType(StatusCodes.Status403Forbidden)]
+        public async Task<ActionResult> DeleteAccount(string username)
         {
-            string username = Environment.UserName;  // vraca username trenutnog korisnika
+            if (User.FindFirstValue(ClaimTypes.Name) != username)
+            {
+                return Forbid();
+            }
 
             var result = await _repository.DeleteMember(username);
 
@@ -83,12 +95,15 @@ namespace IdentityServer.Controllers
         [HttpPut("[action]")]
         [ProducesResponseType(StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status403Forbidden)]
         public async Task<ActionResult> ChangePassword([FromBody] ChangePassword changepassword) {
-            
-            
-            string username = Environment.UserName; 
 
-            var result = await _repository.ChangePassword(username, changepassword.password, changepassword.newPassword);
+            if (User.FindFirstValue(ClaimTypes.Name) != changepassword.username)
+            {
+                return Forbid();
+            }
+
+            var result = await _repository.ChangePassword(changepassword.username, changepassword.password, changepassword.newPassword);
 
             if (!result.Succeeded)
             {
@@ -102,32 +117,6 @@ namespace IdentityServer.Controllers
 
             return StatusCode(StatusCodes.Status200OK);
             
-        }
-
-        [Authorize]
-        [HttpPut("[action]")]
-        [ProducesResponseType(StatusCodes.Status200OK)]
-        [ProducesResponseType(StatusCodes.Status400BadRequest)]
-        public async Task<ActionResult> ChangeUserName([FromBody] ChangeUserName changeusername)
-        {
-
-
-            string username = Environment.UserName;
-
-            var result = await _repository.ChangeUserName(username, changeusername.newUserName);
-
-            if (!result.Succeeded)
-            {
-                foreach (var error in result.Errors)
-                {
-                    ModelState.TryAddModelError(error.Code, error.Description);
-                }
-
-                return BadRequest(ModelState);
-            }
-
-            return StatusCode(StatusCodes.Status200OK);
-
         }
